break apart layouts, add tests for other provided payloads
