+++ sdk/definitions/src/index.ts
@@ -1,10 +1,4 @@
 export * from "./ntt.js";
-export type {
-  TrimmedAmount,
-  Prefix,
-  NativeTokenTransfer,
-  TransceiverMessage,
-  NttManagerMessage,
-  WormholeTransceiverMessage,
-} from "./nttLayout.js";
-export * from "./nttLayout.js";
+
+export * from "./layouts/index.js";
+export type * from "./layouts/index.js";
